requirements.txt
streamlit
pinecone-client
openai
requests
beautifulsoup4